service-roteamento/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
grpcio==1.60.0
grpcio-tools==1.60.0
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
geopy==2.4.0
